# docs/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Bardic server API
  version: 0.2.6
  summary: Contract between the Bardic server and its clients (draft).
  description: |
    Normative, language-neutral contract for Bardic v2. It is derived from
    `docs/PRODUCT-SPEC.md`, not from the prototype. Clients are generated
    from this file; the server must satisfy it. Read `README.md` in this folder
    first: it explains identity headers, errors, money, events, long-running
    work and the place-sync protocol once, so each operation does not repeat them.

    Status: pre-release (0.x). Until 1.0.0, additive changes bump the patch
    version and breaking changes bump the minor version.
  license:
    name: Personal use
    identifier: LicenseRef-Personal
servers:
  - url: http://localhost:8765
    description: Default local address. The server name and address are configurable.
...
paths:
...
  /api/server:
    get:
      operationId: getServer
      tags: [System]
      summary: Server identity and capabilities
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The server.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Server'}
    patch:
      operationId: updateServer
      tags: [System]
      summary: Rename the server
      description: Changing the name never changes the address clients use.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 60}
      responses:
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The updated server.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Server'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
  /api/events:
    get:
      operationId: streamEvents
      tags: [System]
      summary: Live change notifications (Server-Sent Events)
      description: |
        A long-lived `text/event-stream`. Each event is a `Notice` carrying ids
        only; clients re-read the resource. Supports `Last-Event-ID` to resume;
        if the server cannot replay from that id it sends a `resync` notice and
        clients reload what they show.
      x-bardic-cost: none
      parameters:
        - {$ref: '#/components/parameters/Listener'}
        - name: Last-Event-ID
          in: header
          required: false
          schema: {type: string}
      responses:
        '404': {$ref: '#/components/responses/NotFound'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Event stream.
          content:
            text/event-stream:
              schema: {$ref: '#/components/schemas/Notice'}
  /api/devices:
    get:
      operationId: listDevices
      tags: [System]
      summary: Devices that have used this server
      description: Devices register themselves on first request that carries `X-Bardic-Device`.
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Devices, most recently seen first.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeviceList'}
  /api/devices/{device_id}:
    parameters: [{$ref: '#/components/parameters/DeviceId'}]
    patch:
      operationId: updateDevice
      tags: [System]
      summary: Name a device
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string, minLength: 1, maxLength: 60}
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The device.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Device'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ------------------------------------------------------------ Listeners
  /api/listeners:
    get:
      operationId: listListeners
      tags: [Listeners]
      summary: List listeners
      description: Ordered by name. Includes `last_listened_at` so pickers can show recency.
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Listeners.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListenerList'}
    post:
      operationId: createListener
      tags: [Listeners]
      summary: Add a listener
      description: |
        Errors: `name_invalid` (400), `name_taken` (409). The first listener on a
        fresh server is created the same way.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListenerInput'}
      responses:
        '500': {$ref: '#/components/responses/ServerError'}
        '201':
          description: The new listener.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Listener'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/listeners/{listener_id}:
    parameters: [{$ref: '#/components/parameters/ListenerId'}]
    get:
      operationId: getListener
      tags: [Listeners]
      summary: Get a listener
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The listener.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Listener'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      operationId: renameListener
      tags: [Listeners]
      summary: Rename a listener
      description: "Errors: `name_invalid` (400), `name_taken` (409)."
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListenerInput'}
      responses:
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: The listener.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Listener'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
    delete:
      operationId: deleteListener
      tags: [Listeners]
      summary: Delete a listener and their places
      description: |
        Deletes that listener's places, history, finished marks and settings only.
        Books, audio, other listeners and audit records are kept. Errors:
        `last_listener` (409).
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '204': {description: "Deleted."}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/listeners/{listener_id}/impact:
    parameters: [{$ref: '#/components/parameters/ListenerId'}]
    get:
      operationId: getListenerImpact
      tags: [Listeners]
      summary: What deleting a listener would remove
      description: Used by the delete confirmation ("6 books started").
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Counts.
          content:
            application/json:
              schema:
                type: object
                required: [books_started, places, history_entries]
                properties:
                  books_started: {type: integer, minimum: 0}
                  places: {type: integer, minimum: 0}
                  history_entries: {type: integer, minimum: 0}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/listeners/{listener_id}/settings:
    parameters: [{$ref: '#/components/parameters/ListenerId'}]
    get:
      operationId: getListenerSettings
      tags: [Listeners]
      summary: Get a listener's settings
      x-bardic-cost: none
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Settings.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListenerSettings'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: putListenerSettings
      tags: [Listeners]
      summary: Replace a listener's settings
      description: |
        Errors: `voice_not_found` (404) if the default voice does not exist.
        Setting a premium default voice is allowed; pressing play on a new book
        then asks for a plan first (`plan_required`).
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ListenerSettings'}
      responses:
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Settings.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListenerSettings'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '404': {$ref: '#/components/responses/NotFound'}

  # ------------------------------------------------------------ Library
  /api/books:
    get:
      operationId: listBooks
      tags: [Library]
      summary: List books
      description: |
        Books not removed, with the requesting listener's place summary. Filters
        combine with AND. `filter=on_device` is resolved by the client (the server
        does not know what a device holds) and is therefore not a value here.
      x-bardic-cost: none
      parameters:
        - {$ref: '#/components/parameters/Listener'}
        - {name: q, in: query, schema: {type: string}, description: "Matches title, author and series name."}
        - name: filter
          in: query
          schema: {type: string, enum: [all, in_progress, not_started, finished]}
        - name: sort
          in: query
          schema: {type: string, enum: [recent, title, author, added], default: recent}
        - {name: include_removed, in: query, schema: {type: boolean, default: false}}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/After'}
      responses:
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: A page of books.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BookPage'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
  /api/books/duplicates:
    get:
      operationId: findDuplicateBooks
      tags: [Library]
      summary: Check a file against the library before uploading it
      description: |
        `sha256` is the lowercase hex SHA-256 of the file. Includes removed books.
        Never blocks an import; it only informs.
      x-bardic-cost: none
      parameters:
        - {name: sha256, in: query, required: true, schema: {type: string, pattern: '^[0-9a-f]{64}$'}}
        - {name: title, in: query, schema: {type: string}}
        - {name: author, in: query, schema: {type: string}}
      responses:
        '500': {$ref: '#/components/responses/ServerError'}
        '200':
          description: Matches.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DuplicateBooks'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
...
  /api/books/sample:
    post:
      operationId: createSampleBook
      tags: [Library]
      summary: Add the built-in sample book
      description: Adds an original short story. Free, instant, contacts no provider.
      x-bardic-cost: none
      parameters: [{$ref: '#/components/parameters/Device'}]
      responses:
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '500': {$ref: '#/components/responses/ServerError'}
        '201':
          description: The book.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Book'}
...
components:
  parameters:
...
    Device:
      name: X-Bardic-Device
      in: header
      required: true
      description: Client-generated device id (UUID). Registers the device on first use and is recorded in the audit log.
      schema: {type: string}
...
    DeviceId: {name: device_id, in: path, required: true, schema: {type: string}}
...
  responses:
    InvalidRequest:
      description: The request is malformed or breaks a rule; `code` says which.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    ServerError:
      description: An unexpected failure on the server. Nothing you need to undo was changed; `retryable` is true.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
...
    Server:
      type: object
      required: [id, name, version, api_version, max_upload_bytes, free_bytes, time]
      properties:
        id: {type: string, description: "Stable id of this server's data folder."}
        name: {type: string}
        version: {type: string}
        api_version: {type: string}
        max_upload_bytes: {type: integer}
        free_bytes: {type: [integer, 'null'], description: "Free space on the data volume; null if unknown."}
        time: {type: string, format: date-time}
...